fix: removes `getattr` on nonexistant `_entityDict`
* Adds some type annotations
* Converts to Poetry project
* **HARDCODES RANGE FILTER**

Tableau's OData connector does not seem to allow setting a filter prior
to extracting a data set. Influx will throw an error for data sets that
may not be finite if no range filter is given. This can be worked around
by hardcoding a range expression into the adapter.

A better solution might be to make the range filter a configurable
value, or compute the filter from the existing
`data_availability_timerange` option.